Generalized parse code
