+++ ProdirectScraper/configuration.ini
@@ -1,22 +1,26 @@
 [DEFAULTS]
 
-[trainers_spider]
-# use coma separated values from 4 to 12
-size = 7,8
+[general]
 # avaiable currency EUR,USD,GBP
 currency = EUR
-
-[mensclothing_spider]
-currency = EUR
-# use coma separated values with these parameters [One size, ONE-SIZE, S/M, L/XL, S, M, L, XL, XXL]
-size = XL,L/XL
+#number of pieces to display
+pp = 32
 
 [mailer]
 smtp_host =
 mail_from =
-mail_to = zoran.pandovski@gmail.com
+mail_to =
 smtp_user =
 smtp_pass =
 smtp_port =
 smtp_tls =
 smtp_ssl =
+
+[trainers_spider]
+# use coma separated values from 4 to 12
+size = 7,8
+
+[mensclothing_spider]
+# parameters [One size, ONE-SIZE, S/M, L/XL, S, M, L, XL, XXL]
+size = XL,L/XL
+
